refactor: timeformatting 함수 분리
